Created stories file for text.
